Remove LookML view files for outdated raw tables (Recidiviz/recidiviz-data#22167)
## Description of the change

> The set of raw data tables in a state is dynamic and tables can also
be removed (e.g. if a table name is changed). The LookML view generation
script for raw data files will delete view files for raw tables no
longer in the configs list. To ensure that files are not accidentally
deleted, we prompt the user if the directory is not `looker`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | **non-breaking change that adds functionality** |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#21937

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Ryan Guan <rzgn@users.noreply.github.com>
GitOrigin-RevId: ecbb4988a2bbbd4efb67685fa5dc2c880be05ac9