add officer specialized caseload attribute to Outliers views (Recidiviz/recidiviz-data#22431)
This just adds an additional column to the export for supervision
officers that lets us identify officers with a specialized caseload of
some sort. It is not as yet connected to any product feature but this
makes it available to the prototype application, where we are still
exploring how we might make use of it.

GitOrigin-RevId: 1dcbf3f3b154bf0ed444a59145cd85ea13a41c09