[Justice Counts] Update CSG data pull (Recidiviz/recidiviz-data#22372)
## Description of the change

Making a few updates to the data pull! In particular:
* Removing child agencies (since there were just so many of them)
* Adding is_superagency column
* Adding created_at, last_login, and last_update columns. created_at and
last_login are based on Auth0 data, so we are pulling that via their
management API. The last_update column is based on the created_at
timestamp of the agency's datapoints. There are some caveats to the
logic there which I have explained in comments, and I filed tasks for
improvements we can make so that they are more accurate!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6a1aa3562f0f6c2cf1c59c0783fe9574728ce92e