[US_ND] Raw data migrations for `elite_offenderidentifiers` (Recidiviz/recidiviz-data#22450)
## Description of the change

In investigating why this validation was failing Recidiviz/recidiviz-data#18807, we found 21
person_ids in staging and 22 in prod that have more than one external ID
of type US_ND_SID. A large majority of these stemmed from typos in the
raw data `elite_offenderidentifier.IDENTIFIER` field that gets mapped to
`state_person.external_id` with type `US_ND_SID`. Most were identical
except for one digit, two were visibly not SID values, and one was a
date.

There are only three that require further investigation on how to
migrate because both SIDs that are linked to the same `person_id` do
actually exist in `elite_offenders`, meaning they are both real IDs. I
believe these can also be migrated to be correct, but I need to dig a
little more to figure out the correct way to untangle them, and I think
these migrations will bring the validation back down to a soft failure
in the meantime.

This validation started failing around the time that we migrated to v2
mappings, so my best guess is that these data entry errors have always
existed and the change in validation failure level has something to do
with the v2 migration.

Queries and notes dumped [here
](https://docs.google.com/document/d/1Xza17ry-yPY3eSAIDJDy8GVKDuIcM9OXIExsveSzBIY/edit)

I also included a db migration to make sure that all of these IDs are
cleaned up the state database.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18807

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9fcc96b8ebdfd0309f557d649abcf75e88d46bff