[US_TN] Removing hydration of PED in TN  (Recidiviz/recidiviz-data#22394)
## Description of the change

We hydrated PED with ReleaseElgibilityDate in TN in Recidiviz/recidiviz-data#20613 to allow for
state agnostic analyses. In doing so, it revealed an intricacy in the
way TN calculated this RED that led this validation to fail.

According to TN, in many sentence structures both the RED and EXP reduce
at the same
rate with earned reduction credits. Most of the time when you see RED
prior to an effective date, a person
will have been incarcerated for some time and earned quite a few
reduction credits. This will continue to reduce
EXP until EXP matches the month of credit earned and the person would
then be discharged from the sentence.
The RED does not stop reducing once the credit month matches RED due to
the fact there may be consecutive
sentences involved currently or in the future. Until this becomes a
specific issue for analysis, we will bump
the threshold higher to account for this case and refactor eventually
with #TODO(Recidiviz/recidiviz-data#16190)

Because of not being able to isolate this error as opposed to other
sentence date errors in the validation, we are going to remove the
hydration of PED in TN for now and will revisit it as needed in Recidiviz/recidiviz-data#22443.

This PR removes that hydration in the mappings and contains a migration
to wipe the hydrated PEDs in the state dataset in TN.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#20978, Closes Recidiviz/recidiviz-data#20971

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fcf69be675ef6804beadff86f64fd427ee0deb29