[CSG][US_MO][US_PA] Adding validation script for CSG export to ensure data freshness before sending export (Recidiviz/recidiviz-data#20386)
## Description of the change

We've had several issues in recent transfers with CSG where the data we
had was not yet updated for a variety of reasons (raw data we received
was out of date, metric exports didn't finish yet, etc).

This script generate a script to run in BQ to see all the MAX dates in
the various tables we send to CSG and will surface if any of the tables
we expect to have data through the end of the month do not have the
needed data. When they do not, we will not be able to send the export
and instead need to investigate to figure out why we have a data lag.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#20080

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6b3850ad9864084c40002518e6f26aea92338982