Double-write housing unit type and category (Recidiviz/recidiviz-data#22505)
## Description of the change

This re-commits Recidiviz/recidiviz-data#22251, and will be merged _after_ the release candidate
is deployed today (7/21) so that it will go out in a separate prod
deploy from Recidiviz/recidiviz-data#22220

After this schema change is fully implemented,
`StateIncarcerationPeriodHousingUnitType` will become more granular, and
`StateIncarcerationPeriodHousingUnitCategory` will replace it as a
super-type to identify periods of solitary confinement. This change
double-writes the two fields in the one downstream view where
`StateIncarcerationPeriodHousingUnitType` is used to identify periods of
solitary confinement.

I don't think a migration is necessary for this change, since this is
adding a new field reference and not changing any that are already used.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------

|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue | | Type: Feature
| non-breaking change that adds functionality | | Type: Breaking Change
| fix or feature that would cause existing functionality to not work as
expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality | | Type:
Configuration Change | adjusts configuration to achieve some end related
to functionality, development, performance, or security | | Type:
Dependency Upgrade | upgrades a project dependency - these changes are
not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#21778

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 37bc29be60c3e6ed2015c4c6242f57ae12b6377f