Extract `SQLAlchemyEngineManager` / `SQLAlchemyDatabaseKey` dependencies from `local_postgres_helpers` (Recidiviz/recidiviz-data#21691)
## Description of the change
I wanted to use the `local_postgres_helpers` functionality in the
Airflow Pipenv while writing tests for the Airflow monitoring DAG, but
couldn't as `local_postgres_helpers` depended on too many of the
Recidiviz `Pipfile` dependencies. This extracts the SQLAlchemy*
dependencies into a file named `local_persistence_helpers` so that the
postgres-specific functionality can be used in the Airflow pipenv.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#21522

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: da39ee9ed2fe58fa1f5759024affc352d5cc46d2