remove double writes to legacy projects in automatic upload (Recidiviz/recidiviz-data#22553)
## Description of the change

This PR removes double writes to the legacy aggregate buckets buckets so
that we are only reading and writing from aggregate buckets in
`justice-counts-staging` and `justice-counts-production`. Once this
change is out in staging, I will run
[this](https://github.com/Recidiviz/recidiviz-data/blob/0678caedf61040c9a4e0ff14e4a23dd3d2bef2cf/recidiviz/tools/justice_counts/migrate_spreadsheets.py#L54)
script to copy whatever files are not in the new buckets to that
location.

I also added a comment to clarify that agencies with multiple systems
will have a bucket that has a folder that corresponds to each system.

## Related issues

Related to Recidiviz/recidiviz-data#22175

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3901faf283e4b81d8d14836ef5fc8f78281b3ae7