Fixing the trigger rule and success indicator for state branch ends (Recidiviz/recidiviz-data#22601)
## Description of the change

The trigger rule `NONE_FAILED_MIN_ONE_SUCCESS` does not actually wait
until all of the tasks are finished before triggering the task
`state_code_branch_end` and this was causing the
`update_normalized_state` task to be completed before all normalization
pipelines finished running when one has failed. This changes the default
rule to `NONE_SKIPPED` and within the `state_code_branch_end`, we raise
an `AirflowFailException` if any of the upstream task instances have
failed or have a further upstream that has failed (because many of our
task instances are now in task groups). This allows for the state to be
indicated as failure for the entire DAG.

Tested in `experiment-2`. Integration test is following.

## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#22555

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 15ff1b322d05a734b40d32dfed5a5713c62f6dc2