[US_IX] Add reference view for leadership supervisors (Recidiviz/recidiviz-data#22523)
## Description of the change

Add reference view for Idaho leadership people who are also supervisors
so they can be included in the supervisor view of Outliers even when
they're `role_subtype` is not `SUPERVISION_OFFICER_SUPERVISOR`.

To be deprecated in the future when the product views are refactored to
account for this scenario (ticket
[here](https://github.com/Recidiviz/recidiviz-data/issues/22524)).

(I also noticed I accidentally created a reference view script in the
past that I didn't mean to create, so deleting that in this PR as well).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8278d1f70a550888de2dce147c4f56a8ad72e062