Adding projected_end_date to incident outcome (Recidiviz/recidiviz-data#22643)
## Description of the change

Change of adding projected_end_date to the incident outcome schema. This
is helpful for a number of states that provide us with this data so we
can understand if seg stays are ending as intended/ when they were
supposed to end.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 74582bb11d0c1a418fc6f27621149b9592f109bf